+++ examples/cli/src/bufferScan.ts
@@ -23,8 +23,11 @@ const useKey = false
  * scan it.
  *
  * @param fileName - Name of the file to scan
+ * @param tags - List of string to tag a scan
+ * @param pml - Enable predictive machine learning detection
+ * @param feedback: Enable Trend Micro Smart Protection Network's Smart Feedback
  */
-const runBufferScan = async (fileName: string): Promise<void> => {
+const runBufferScan = async (fileName: string, tags: string[], pml: boolean, feedback: boolean): Promise<void> => {
   console.log(`\nScanning buffer '${fileName}'`)
   const amaasGrpcClient = useKey
     ? new AmaasGrpcClient(amaasHostName, key)
@@ -34,7 +37,7 @@ const runBufferScan = async (fileName: string): Promise<void> => {
 
   const buff = readFileSync(fileName)
   await amaasGrpcClient
-    .scanBuffer(fileName, buff)
+    .scanBuffer(fileName, buff, tags, pml, feedback)
     .then(result => {
       console.log(`${JSON.stringify(result)}`)
     })
@@ -47,6 +50,9 @@ const runBufferScan = async (fileName: string): Promise<void> => {
 }
 
 void (async () => {
-  // Examples of using AmaasGrpcClient
-  await runBufferScan('node_modules/@grpc/grpc-js/src/admin.ts')
+  // Examples of using AmaasGrpcClient with tags and TrendX with smart feedback
+  const tags = ['example', 'test']
+  const predictive_machine_learning = true
+  const smart_feedback = true
+  await runBufferScan('node_modules/@grpc/grpc-js/src/admin.ts', tags, predictive_machine_learning, smart_feedback)
 })()