+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.1.0 - 2024-04-03
+
+- Update protos
+- Enable PML (Predictive Machine Learning) detection and smart feedback
+- Enable bulk mode
+- Enable India region
+- Support for scanning large files (over 2GB)
+
 ## 1.0.3 - 2024-01-16
 
 - Create static sha test file for unit test